+++ CHANGELOG.rst
@@ -17,6 +17,11 @@ Changed
 `2.0b2` - 2021-06-15
 --------------------
 
+Added
+~~~~~~~
+
+* Add option to disable dashboard auto updating as a user and sitewide.
+
 Changed
 ~~~~~~~